made service category a numeric

+++ General_Resources/Views/Materialized/2016 Material Girl Equivalents/fy2016_districts_c2_funding_matr.sql
@@ -218,7 +218,7 @@ from (
     left join (
       select *
       from fy2016.line_items
-      where service_category = '2.0'
+      where service_category::numeric = '2.0'
     ) c2_li_2016
     on c2_li_2016.id = ad_2016.line_item_id
     left join (
@@ -297,7 +297,7 @@ from (
       left join (
         select *
         from fy2016.line_items
-        where service_category = '2.0'
+        where service_category::numeric = '2.0'
       ) c2_li_2016
       on c2_li_2016.id = ad_2016.line_item_id
       left join (
